refactor: split vectorised path draws into sim package
